Add files via upload
This program needs a lot of work. It is not modularized very well. The flood_fill function is ridiculously long, and it doesn't do what the traditional flood_fill function should do. However, this program does work, and a user can play the game of Minesweeper. I did this program a year ago. I would very much like to redo it soon.